Refactor debug and InfluxDB connection logic
Changed debugMessage to use uint8_t for messageLevel across files. Updated InfluxDB connection logic to use a timeout-based approach instead of fixed retry attempts. Refactored RSSI handling to use hardwareData.rssi and improved configuration and secrets templates for clarity and consistency.

+++ post_mqtt.cpp
@@ -15,7 +15,7 @@
   extern PubSubClient mqtt;
 
   // Shared helper function
-  extern void debugMessage(String messageText, int messageLevel);
+  extern void debugMessage(String messageText, uint8_t messageLevel);
 
   #ifdef HASSIO_MQTT
     extern void hassio_mqtt_setup();
